chore(travis): fix travis ci build script; added license

+++ grunt-tasks/options/shell.js
@@ -1,3 +1,4 @@
+/*jshint node:true*/
 module.exports = {
     //We use %version% and evluate it at run-time, because <%= pkg.version %>
     //is only evaluated once
@@ -21,10 +22,15 @@ module.exports = {
         ].join(' && ')
     },
     hooks: {
-        command: [
-            'rm -f .git/hooks/jshint .git/hooks/validate-commit-msg',
-            'cp grunt-tasks/git-hooks/* .git/hooks/',
-            'chmod u+x .git/hooks/jshint .git/hooks/validate-commit-msg'
-        ].join(' && ')
+        command: function () {
+            // don't run the command if we're in Travis CI
+            if (!process.env.TRAVIS) {
+                return [
+                    'rm -f .git/hooks/jshint .git/hooks/validate-commit-msg',
+                    'cp grunt-tasks/git-hooks/* .git/hooks/',
+                    'chmod u+x .git/hooks/jshint .git/hooks/validate-commit-msg'
+                ].join(' && ');
+            }
+        }
     }
 };
